Versione 2.0

+++ alex_tree.c
@@ -66,7 +66,7 @@ static void TreeGetLevelValuesRec(const Node* n, const ElemType** ris, size_t le
 const ElemType** TreeGetLevelValues(const Node* n, size_t level)
 {
 	if (n == NULL) { return NULL; }
-	const ElemType** ris = calloc(1 << level, sizeof(ElemType*));
+	const ElemType** ris = calloc((size_t)1 << level, sizeof(ElemType*));
 	size_t index = 0;
 	TreeGetLevelValuesRec(n, ris, level, 0, &index);
 	return ris;
